Do autoconf wrangling in a subshell
Avoid polluting the main configure phase environment with the changes
made to the flagsArray array and to PATH to build the configure file.
This probably makes zero difference most of the time, but reduces the
risk that something that would break the regular Nixpkgs builds gets
missed because it still works here.